Fix "new table" problem in module function, Add exporting a module to specified lua table
Table newed in module function by temporary handle can't be returned to lua, because temporary handle will clean the new table while destruction which happened before MakeRet.
Sometimes we just need some functions in C++, then we define a module who wraps these functions, but we don't want the module to be a global table in lua, so just export the utility module to a local lua table.